+++ both/components/LineChart/LineChart.jsx
@@ -1,28 +1,46 @@
 LineChart = React.createClass({
 
+//	                data: [65, 59, 80, 81, 56, 55]
+//	                data: [28, 48, 40, 19, 86, 27]
+
+    mixins: [ReactMeteorData],
+
+    getMeteorData() {
+    	return {
+    		facility:Session.get('selectedFacility')
+    	}
+    },
+
+    getInitialState() {
+    	return {
+    		initialised:false
+    	}
+    },
+	
+
 	initChart() {
 	    var lineData = {
-	        labels: ["January", "February", "March", "April", "May", "June", "July"],
+	        labels: ["January", "February", "March", "April", "May", "June"],
 	        datasets: [
 	            {
-	                label: "Example dataset",
+	                label: "Open",
 	                fillColor: "rgba(220,220,220,0.5)",
 	                strokeColor: "rgba(220,220,220,1)",
 	                pointColor: "rgba(220,220,220,1)",
 	                pointStrokeColor: "#fff",
 	                pointHighlightFill: "#fff",
 	                pointHighlightStroke: "rgba(220,220,220,1)",
-	                data: [65, 59, 80, 81, 56, 55, 40]
+	                data: [0,0,0,0,0,0]
 	            },
 	            {
-	                label: "Example dataset",
+	                label: "Closed",
 	                fillColor: "rgba(26,179,148,0.5)",
 	                strokeColor: "rgba(26,179,148,0.7)",
 	                pointColor: "rgba(26,179,148,1)",
 	                pointStrokeColor: "#fff",
 	                pointHighlightFill: "#fff",
 	                pointHighlightStroke: "rgba(26,179,148,1)",
-	                data: [28, 48, 40, 19, 86, 27, 90]
+	                data: [0,0,0,0,0,0]
 	            }
 	        ]
 	    };
@@ -41,18 +59,51 @@ LineChart = React.createClass({
 	        datasetStrokeWidth: 2,
 	        datasetFill: true,
 	        responsive: true,
+			legendTemplate : "<ul class=\"chart-legend\"><% for (var i=0; i<datasets.length; i++){%><li><span style=\"background-color:<%=datasets[i].strokeColor%>\"></span><%if(datasets[i].label){%><%=datasets[i].label%><%}%></li><%}%></ul>"
 	    };
 
-	    var ctx = document.getElementById("lineChart").getContext("2d");
-	    var myNewChart = new Chart(ctx).Line(lineData, lineOptions);
-   	},
+	    var ctx = document.getElementById("line-chart").getContext("2d");
+	    this.chart = new Chart(ctx).Line(lineData, lineOptions);
+		this.legend = this.chart.generateLegend();
+  		$('#line-chart-wrapper').append(this.legend); 
+
+  		this.setState({
+  			initialised:true
+  		});
+	},
+
+	createRandomSet(length,max) {
+		var set = [];
+		for(var i=0;i<length;i++) {
+			set.push(Math.floor(Math.random()*max));
+		}
+		return set;
+	},
+
+	updateData() {
+		if(!this.state.initialised)
+			return;
+        var data = {
+        	closed:this.createRandomSet(6,10),
+        	open:this.createRandomSet(6,10)
+        }
+        for(var i=0;i<6;i++) {
+	        this.chart.datasets[0].points[i].value = data.closed[i];
+	        this.chart.datasets[1].points[i].value = data.open[i];
+        }
+        this.chart.update();
+	},
+
 	componentDidMount() {
         this.initChart();
 	},
+
+
 	render() {
+		this.updateData();
 	    return (
-	    	<div>
-	    		<canvas id="lineChart"></canvas>
+	    	<div id="line-chart-wrapper">
+	    		<canvas id="line-chart"></canvas>
 	    	</div>
 	    )
 	}