+++ both/views/pages/PageDashboard.jsx
@@ -9,11 +9,8 @@ PageDashboard = React.createClass({
         <div className="row">
             <div className="col-lg-6 col-md-6 col-sm-12">
 	            <Box title="Work Requests">
-	            	<div className="btn-group">
-	            		<button className="btn btn-white"><i className="fa fa-th-list"></i> Day</button>
-	            		<button className="btn btn-white"><i className="fa fa-th-large"></i> Week</button>
-	            		<button className="btn btn-white"><i className="fa fa-th"></i> Month</button>
-	            		<button className="btn btn-white"><i className="fa fa-th"></i> 3 Month</button>
+	            	<div className="btn-group pull-right">
+	            		<button className="btn btn-white"><i className="fa fa-th"></i> 6 Month</button>
 	            	</div>
 	            	<div className="row">
 	            		<div className="col-md-12">
@@ -21,14 +18,11 @@ PageDashboard = React.createClass({
 	            		</div>
 	            	</div>
 	            </Box>
-	            <Box title="Work Requests">
-	            	<div className="btn-group">
-	            		<button className="btn btn-white"><i className="fa fa-th-list"></i> Day</button>
-	            		<button className="btn btn-white"><i className="fa fa-th-large"></i> Week</button>
+	            <Box title="Building Overview">
+	            	<div className="btn-group pull-right">
 	            		<button className="btn btn-white"><i className="fa fa-th"></i> Month</button>
-	            		<button className="btn btn-white"><i className="fa fa-th"></i> 3 Month</button>
 	            	</div>
-	            	<div className="row" style={{textAlign:"center"}}>
+	            	<div className="row" style={{textAlign:"center",clear:"both"}}>
 						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
 							<ProgressArc title="Requests" value={{thisMonth:15,lastMonth:20}}/>
 				        </div>
@@ -48,25 +42,22 @@ PageDashboard = React.createClass({
 							<ProgressArc title="Supplier Reviews" value={{thisMonth:5,lastMonth:8}} color="orange"/>
 				        </div>
 						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
-							<ProgressArc title="Open Requests" value={{thisMonth:5,lastMonth:7}} color="green"/>
+							<ProgressArc title="Contracts Expiring" value={{thisMonth:5,lastMonth:7}} color="green"/>
 				        </div>
 						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
-							<ProgressArc title="Open Orders" value={{thisMonth:5,lastMonth:8}} color="orange"/>
+							<ProgressArc title="Insurance Expiring" value={{thisMonth:5,lastMonth:8}} color="orange"/>
 				        </div>
 		            </div>
 	            </Box>
 	        </div>
             <div className="col-lg-6 col-md-6 col-sm-12">
-	            <Box title="Active Building Compliance">
-	            	<div className="btn-group">
-	            		<button className="btn btn-white"><i className="fa fa-th-list"></i> Day</button>
-	            		<button className="btn btn-white"><i className="fa fa-th-large"></i> Week</button>
-	            		<button className="btn btn-white"><i className="fa fa-th"></i> Month</button>
-	            		<button className="btn btn-white"><i className="fa fa-th"></i> 3 Month</button>
+	            <Box title="Non-Compliant Items">
+	            	<div className="btn-group pull-right">
+	            		<button className="btn btn-white"><i className="fa fa-th"></i> 6 Month</button>
 	            	</div>
 	            	<BarChart />
 	            </Box>
-	            <Box title="Preventative Maintenence Planner">
+	            <Box title="Preventative Maintenance Planner">
 	            	<Calendar />
 	            </Box>
 	        </div>